requirements.txt
fastapi
uvicorn
pytesseract
pillow
opencv-python-headless
requests
numpy
python-dotenv
